legacy/requirements.txt
playwright>=1.57.0
pytest>=8.0.0
psycopg2-binary>=2.9
psycopg2-binary>=2.9
